Add loratx service and commands

+++ core/services/lora_tx_service.cpp
@@ -40,7 +40,7 @@ unsigned int LoRaPacketBuilder::convert_longitude(double x) {
 }
 
 unsigned int LoRaPacketBuilder::convert_heading(double x) {
-	return x / 1.42f;
+	return (unsigned int)(x / 1.42);
 }
 
 unsigned int LoRaPacketBuilder::convert_altitude(double x) {
